Basic port for e2k(elbrus) architecture.
Basic port for e2k (minidump, microdump, stackwalker, md2core).
Main problem was to save 3 stacks instead of 1 (user, procedure and chain) and to walk through them using chain file.
Only a small part of the registers is saved in the context needed for debug, frame registers can be obtained from procedure stack, but they are not needed in stackwalker, it is much easier to work with them through core file, where they are available.